added possible correction for relative paths in links
I added a hacky but hopefully effective way of correcting relative paths to files inside the module root. It uses regular expressions with grepl and gsub to replace "](" in "[text](link)" and ignoring cases where the link starts with "http" (another alternative is use ".htm" assuming that the whole line doesn't contain this string elsewhere. Maybe this can be improve to narrow down the exceptions without limiting the content of the line too much. Note that we just need to make sure that all html links in READMEs really start with https://... I also left commented out the one-liner option, which might be feasible if someone finds the way of writing both the pattern and exception in the same regular expression.

Co-Authored-By: Clemens Schmid <clemens@nevrome.de>